Struct file_sync::FileSync

pub struct FileSync<T>where
    T: Serialize + DeserializeOwned,{
    pub pretty: bool,
    /* private fields */
}

Note: Methods that take a &mut self and return a Result might cause de-sync between the internal data and the file if the Result is an Err

Fields
pretty: bool

Specifies if when writing to the file if serde_json::to_writer_pretty will be used instead of serde_json::to_writer

Implementations

impl<T> FileSync<T>where T: Serialize + DeserializeOwned,

pub fn new(fp: &Path, data: T, pretty: bool) -> Result<Self, FileSyncError<'_>>

Creates a new FileSync type syncing a file with the path fp and data

pretty determines if it will use serde_json::to_writer_pretty instead of serde_json::to_writer

Errors

Will return an error if a file at that path already exists

Will return an error if the creating the File returns an error

Will return an error if serde_json::to_writer/serde_json::to_writer_pretty returns an error

pub fn load(fp: &Path, pretty: bool) -> Result<Self, FileSyncError<'_>>

Creates a new FileSync type loading and syncing data from an already existing file

pretty determines if iet will use serde_json::to_writer_pretty instead of serde_json::to_writer

Errors

Will return an error if the creating the File returns an error

Will return an error if serde_json::from_reader returns an error

pub fn load_or_new( fp: &Path, data: T, pretty: bool ) -> Result<Self, FileSyncError<'_>>

Creates a new FileSync type loading and syncing data from an already existing file, or creating a new one if the file doesn’t exist

pretty determines if iet will use serde_json::to_writer_pretty instead of serde_json::to_writer

Errors

Will return an error if the creating the File returns an error

Will return an error if serde_json::to_writer/serde_json::to_writer_pretty returns an error

Will return an error if serde_json::from_reader returns an error

pub fn set(&mut self, data: T) -> Result<(), FileSyncError<'_>>

Sets the value of the stored data

Errors

Returns an error if serde_json::to_writer/serde_json::to_writer_pretty returns an error Returns an error if it fails to clear the file

pub fn get(&self) -> &T

Returns an immutable reference to the stored data

pub fn modify<F>(&mut self, f: F) -> Result<(), FileSyncError<'_>>where F: FnOnce(&mut T),

Modifies data and syncs the modified data to the file given a Fn(&mut T)

Errors

Returns an error if serde_json::to_writer/serde_json::to_writer_pretty returns an error Returns an error if it fails to clear the file